+++ src/commands/export.ts
@@ -3,16 +3,15 @@ import * as path from 'path';
 
 // ok
 import * as models from '../models';
-import { CommandBase } from './base';
-import { ExtensionCommandError } from './extensionCommandError';
 
 /**
  * Export command.
  */
-export class Export extends CommandBase {
-  constructor(private vscodeManager: models.IVSCodeManager, bookmarkManager: models.IBookmarkManager) {
-    super(bookmarkManager);
-  }
+export class Export implements models.ICommand {
+  constructor(
+    private vscodeManager: models.IVSCodeManager,
+    private bookmarkManager: models.IBookmarkManager,
+  ) {}
 
   /**
    * Return the command name.
@@ -28,7 +27,7 @@ export class Export extends CommandBase {
   public execute(
     _execArgs: models.IVSCodeExecutableArguments,
     configManager: models.IConfigManager,
-    _bookMarkManager: models.IBookmarkManager
+    _bookMarkManager: models.IBookmarkManager,
   ): void {
     // validate cofiguration.
     var [ok, reason] = configManager.validate();
@@ -41,14 +40,12 @@ export class Export extends CommandBase {
     var bookmarksInfo: models.IBookmarksInfo;
     try {
       let fullPath = configManager.defaultBookmarkFullPath();
-      bookmarksInfo = this.loadBookmarksInfo(fullPath ? fullPath : '');
+      bookmarksInfo = this.bookmarkManager.loadBookmarksInfo(
+        fullPath ? fullPath : '',
+      );
     } catch (e) {
-      if (e instanceof ExtensionCommandError) {
-        this.vscodeManager.window.showWarningMessage(e.message);
-        return;
-      } else {
-        throw e;
-      }
+      this.vscodeManager.window.showWarningMessage(e.message);
+      return;
     }
 
     try {
@@ -58,10 +55,16 @@ export class Export extends CommandBase {
 
       if (currentRootFolder) {
         var exportPath = path.join(currentRootFolder, 'export-bookmarks.json');
-        fs.writeFileSync(exportPath, JSON.stringify(bookmarksInfo), { encoding: 'utf-8' });
-        this.vscodeManager.window.showInformationMessage(`Export Success. (${exportPath})`);
+        fs.writeFileSync(exportPath, JSON.stringify(bookmarksInfo), {
+          encoding: 'utf-8',
+        });
+        this.vscodeManager.window.showInformationMessage(
+          `Export Success. (${exportPath})`,
+        );
       } else {
-        this.vscodeManager.window.showWarningMessage(`Directory has not been opened.`);
+        this.vscodeManager.window.showWarningMessage(
+          `Directory has not been opened.`,
+        );
       }
     } catch (e) {
       this.vscodeManager.window.showErrorMessage(e.message);