+++ src/commands/show.ts
@@ -1,18 +1,18 @@
 import * as open from 'open';
+import * as path from 'path';
 import * as fileutils from '../utils/file';
 
 // ok
 import * as models from '../models';
-import { CommandBase } from './base';
-import { ExtensionCommandError } from './extensionCommandError';
 
 /**
  * Export command.
  */
-export class Show extends CommandBase {
-  constructor(private vscodeManager: models.IVSCodeManager, bookmarkManager: models.IBookmarkManager) {
-    super(bookmarkManager);
-  }
+export class Show implements models.ICommand {
+  constructor(
+    private vscodeManager: models.IVSCodeManager,
+    private bookmarkManager: models.IBookmarkManager,
+  ) { }
 
   /**
    * Return the command name.
@@ -40,45 +40,56 @@ export class Show extends CommandBase {
     // load file
     var bookmarksInfo: models.IBookmarksInfo;
     try {
+      // global
       let fullPath = configManager.defaultBookmarkFullPath();
-      bookmarksInfo = this.loadBookmarksInfo(fullPath ? fullPath : '');
-    } catch (e) {
-      if (e instanceof ExtensionCommandError) {
-        this.vscodeManager.window.showWarningMessage(e.message);
-        return;
-      } else {
-        throw e;
-      }
-    }
+      bookmarksInfo = this.bookmarkManager.loadBookmarksInfo(
+        fullPath ? fullPath : '',
+      );
 
-    var concatBk = this.concatBookmark(
-      bookmarksInfo.fileBookmarks,
-      bookmarksInfo.folderBookmarks,
-      bookmarksInfo.urlBookmarks,
-    );
-    var items = concatBk.map<models.IBookmarkLabel>(b => this.bookmarkManager.createBookmarkLabel(b));
-    if (items.length === 0) {
-      this.vscodeManager.window.showWarningMessage('Bookmark has not been registered.');
+      // workspace
+      let currentRootFolder = this.vscodeManager.workspace.workspaceFolders
+        ? this.vscodeManager.workspace.workspaceFolders[0].uri.path
+        : undefined;
+      if (currentRootFolder) {
+        let wkBookmarksPath = path.join(currentRootFolder, '.vscode', configManager.defaultFileName());
+        let wkBookmarksInfo = this.bookmarkManager.loadBookmarksInfo(wkBookmarksPath);
+        this.bookmarkManager.concatBookmarksInfo(bookmarksInfo, wkBookmarksInfo);
+      }
+    } catch (e) {
+      this.vscodeManager.window.showWarningMessage(e.message);
       return;
     }
-    this.vscodeManager.window.showQuickPick(items, { matchOnDescription: true, matchOnDetail: true }).then(item => {
-      if (!item) {
-        return;
-      }
-      switch (item.type) {
-        case 'file':
-          this.showFile(item.description);
-          break;
-        case 'folder':
-          this.showFolder(configManager, item.description);
-          break;
-        case 'url':
-          this.showUrl(item.description);
-          break;
-        default:
-          break;
-      }
-    });
+
+    var concatBk = this.bookmarkManager.sortAndConcatBookmark(bookmarksInfo);
+    var items = concatBk.map<models.IBookmarkLabel>(b =>
+      this.bookmarkManager.createBookmarkLabel(b),
+    );
+    if (items.length === 0) {
+      this.vscodeManager.window.showWarningMessage(
+        'Bookmark has not been registered.',
+      );
+      return;
+    }
+    this.vscodeManager.window
+      .showQuickPick(items, { matchOnDescription: true, matchOnDetail: true })
+      .then(item => {
+        if (!item) {
+          return;
+        }
+        switch (item.type) {
+          case 'file':
+            this.showFile(item.description);
+            break;
+          case 'folder':
+            this.showFolder(configManager, item.description);
+            break;
+          case 'url':
+            this.showUrl(item.description);
+            break;
+          default:
+            break;
+        }
+      });
   }
 
   /**
@@ -87,10 +98,13 @@ export class Show extends CommandBase {
    */
   private showFile(description: string | undefined) {
     if (description) {
-      var path = fileutils.resolveHome(description);
-      this.vscodeManager.window.showTextDocument(this.vscodeManager.urlHelper.file(path), {
-        preview: false,
-      });
+      var openPath = fileutils.resolveHome(description);
+      this.vscodeManager.window.showTextDocument(
+        this.vscodeManager.urlHelper.file(openPath),
+        {
+          preview: false,
+        },
+      );
     }
   }
 
@@ -99,17 +113,29 @@ export class Show extends CommandBase {
    * @param configManager Fuzzy Bookmark configuration manager.
    * @param description bookmark description.
    */
-  private showFolder(configManager: models.IConfigManager, description: string | undefined) {
+  private showFolder(
+    configManager: models.IConfigManager,
+    description: string | undefined,
+  ) {
     if (description) {
-      var path = fileutils.resolveHome(description);
+      var openPath = fileutils.resolveHome(description);
       switch (configManager.directoryOpenType()) {
         case 'terminal':
           // eslint-disable-next-line max-len
           // refs: https://github.com/microsoft/vscode/blob/94c9ea46838a9a619aeafb7e8afd1170c967bb55/src/vs/workbench/contrib/externalTerminal/browser/externalTerminal.contribution.ts#L30-L83
-          this.vscodeManager.commands.executeCommand('openInTerminal', this.vscodeManager.urlHelper.file(path));
+          this.vscodeManager.commands.executeCommand(
+            'openInTerminal',
+            this.vscodeManager.urlHelper.file(openPath),
+          );
           break;
         case 'explorer':
-          open(path);
+          open(openPath);
+          break;
+        case 'window':
+          this.vscodeManager.commands.executeCommand(
+            'vscode.openFolder',
+            this.vscodeManager.urlHelper.file(openPath),
+          );
           break;
         default:
           break;