+++ src/configuration/configManager.ts
@@ -8,6 +8,7 @@ import * as models from '../models';
 const CONFIG_CATEGORY = 'fzb';
 
 export class ConfigManager implements models.IConfigManager {
+  private filename: string = 'bookmarks.json';
   private config: models.IVSCodeWorkspaceConfiguration;
   constructor(private vscodeManager: models.IVSCodeManager) {
     if (!vscodeManager) {
@@ -18,19 +19,19 @@ export class ConfigManager implements models.IConfigManager {
   }
 
   /**
-   * Returns the default directory path to save the bookmark.
-   * @returns defaultBookmarkDir
+   * Returns the directory path to save the bookmark.
+   * @returns saveBookmarkDir
    */
-  public defaultDir(): string | undefined {
-    return this.config.get<string>(models.ConfigurationKeys.defaultBookmarkDir);
+  public saveDirectoryPath(): string | undefined {
+    return this.config.get<string>(models.ConfigurationKeys.saveBookmarkDir);
   }
 
   /**
    * Returns the default file name where the bookmarks are saved.
    * @returns defaultFileName
    */
-  public defaultFileName(): string | undefined {
-    return this.config.get<string>(models.ConfigurationKeys.defaultBookmarkFileName);
+  public defaultFileName(): string {
+    return this.filename;
   }
 
   /**
@@ -46,7 +47,7 @@ export class ConfigManager implements models.IConfigManager {
    * @returns bookmarkFullPath
    */
   public defaultBookmarkFullPath(): string | undefined {
-    var dir = this.defaultDir();
+    var dir = this.saveDirectoryPath();
     var file = this.defaultFileName();
     if (dir && file) {
       return path.join(dir, file);
@@ -59,28 +60,22 @@ export class ConfigManager implements models.IConfigManager {
    */
   public validate(): [boolean, models.IInValidReason] {
     // Checking the setting values
-    if (!this.defaultDir() || !this.defaultDir()?.trim()) {
+    if (!this.saveDirectoryPath() || !this.saveDirectoryPath()?.trim()) {
       return [
         false,
         {
           // eslint-disable-next-line max-len
-          error: `Set the directory path where Bookmarks will be stored to "${CONFIG_CATEGORY + models.ConfigurationKeys.defaultBookmarkDir}" .`,
+          error: `Set the directory path where Bookmarks will be stored to "${CONFIG_CATEGORY + models.ConfigurationKeys.saveBookmarkDir}" .`,
         },
       ];
     }
-    if (!this.defaultFileName() || !this.defaultFileName()?.trim()) {
-      return [
-        false,
-        // eslint-disable-next-line max-len
-        { error: `Set the file name for saving bookmarks to "${CONFIG_CATEGORY + models.ConfigurationKeys.defaultBookmarkFileName}" .` }];
-    }
 
-    if (!fs.existsSync(fileutils.resolveHome(this.defaultDir()))) {
+    if (!fs.existsSync(fileutils.resolveHome(this.saveDirectoryPath()))) {
       return [
         false,
         {
           // eslint-disable-next-line max-len
-          error: `The configured folder(${this.defaultDir()}) does not exist. Execute "FzB: Setup Fuzzy Bookmarks" or create a folder.`,
+          error: `The configured folder(${this.saveDirectoryPath()}) does not exist. Execute "FzB: Setup Fuzzy Bookmarks" or create a folder.`,
         },
       ];
     }