+++ src/models/bookmark/index.ts
@@ -2,3 +2,4 @@ export * from './bookmarkManager';
 export * from './bookmark';
 export * from './bookmarksInfo';
 export * from './bookmarkLabel';
+export * from './bookmarkError';