+++ src/models/configuration/configManager.ts
@@ -2,8 +2,8 @@ import { DirectoryOpenType } from '../contributions';
 import { IInValidReason } from './invalidReason';
 
 export interface IConfigManager {
-  defaultDir(): string | undefined;
-  defaultFileName(): string | undefined;
+  saveDirectoryPath(): string | undefined;
+  defaultFileName(): string;
   directoryOpenType(): DirectoryOpenType | undefined;
   defaultBookmarkFullPath(): string | undefined;
   validate(): [boolean, IInValidReason];