+++ src/models/contributions/configuration.ts
@@ -1,0 +1,6 @@
+export type DirectoryOpenType = 'terminal' | 'explorer' | 'window';
+
+export enum ConfigurationKeys {
+  saveBookmarkDir = 'saveBookmarkDir',
+  directoryOpenType = 'directoryOpenType',
+}