+++ src/models/contributions/configurationKeys.ts
@@ -1,7 +1,0 @@
-export type DirectoryOpenType = 'terminal' | 'explorer';
-
-export enum ConfigurationKeys {
-  defaultBookmarkDir = 'defaultBookmarkDir',
-  defaultBookmarkFileName = 'defaultBookmarkFileName',
-  directoryOpenType = 'directoryOpenType',
-}