+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Change Log
 
+## 0.4.0 - 10 May 2021
+
+## Information
+
+The key name in Configuration has been changed.
+Please be careful if you are configuring individual settings.
+
+- bef: `fzb.defaultBookmarkDir` → aft: `fzb.saveBookmarkDir`
+- bef: `fzb.defaultBookmarkFileName` → delete.
+
+### [New or Update Feature]
+
+- [#47](https://github.com/atEaE/fuzzy-bookmarks/issues/47) Open a new window when a folder is selected with the Show command.
+- [#3](https://github.com/atEaE/fuzzy-bookmarks/issues/3) Japanese support for Command characters
+- [#49](https://github.com/atEaE/fuzzy-bookmarks/issues/49) I want to place a Workspace-only Bookmark.json in the .vscode
+
+### [Bugfix]
+
+- none
+
+### [Documentation]
+
+- [#7](https://github.com/atEaE/fuzzy-bookmarks/issues/7) Set the icon.
+
+***
+
 ## 0.3.1 - 6 May 2021
 
 Release notes: [#41](https://github.com/atEaE/fuzzy-bookmarks/issues/41)
@@ -15,6 +41,8 @@ Release notes: [#41](https://github.com/atEaE/fuzzy-bookmarks/issues/41)
 ### [Documentation]
 
 - none
+
+***
 
 ## 0.3.0 - 5 May 2021
 
@@ -34,6 +62,8 @@ Release notes: [#27](https://github.com/atEaE/fuzzy-bookmarks/issues/27)
 
 - [#31](https://github.com/atEaE/fuzzy-bookmarks/issues/31) There is no badge on the README.md
 - [#28](https://github.com/atEaE/fuzzy-bookmarks/issues/28) want to divide the PR template into "Merge" and "Release".
+
+***
 
 ## 0.2.0 - 29 April 2021
 
@@ -55,6 +85,8 @@ Release notes: [#12](https://github.com/atEaE/fuzzy-bookmarks/issues/12)
 - [#13](https://github.com/atEaE/fuzzy-bookmarks/issues/13) Workflow does not stop when lint error is raised
 - [#20](https://github.com/atEaE/fuzzy-bookmarks/issues/20) The "PR template" does not have a field for meta issues.
 - [#24](https://github.com/atEaE/fuzzy-bookmarks/issues/24) Add a column for "Estimated release date" to the ReleaseIssue.
+
+***
 
 ## 0.1.0 - 27 April 2021